Rename data folder to lib
