+++ src/lib/getProducts.js
